pyproject.toml
[project]
name = "soke"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.38",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "plotly>=5.22",
  "requests>=2.32",
  "pdfplumber>=0.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
